==> notebooks/04_split_read_counting.ipynb <==
Loads in the csv file containing filtered clusters

import pandas as pd

CLUSTER_CSV = "../results/filtered_NUMT_candidates.csv"

clusters = pd.read_csv(CLUSTER_CSV)
clusters["num_split_reads"] = 0

print(f"Loaded {len(clusters)} clusters")

Builds a map that maps each read to its cluster

# Build a read -> cluster mapping

CLIPPED_READS_CLUSTER_CSV = "../results/clipped_reads_with_cluster.csv"
clipped_df = pd.read_csv(CLIPPED_READS_CLUSTER_CSV)

read_cluster_map = dict(zip(clipped_df["read_name"], clipped_df["cluster_idx"]))

for cluster_idx in clipped_df["cluster_idx"]:
    clusters.at[cluster_idx, "num_split_reads"] += 1

print(clusters.head())

Adds cluster index information

import pysam

MT_BAM = "../results/clipped_vs_mtDNA_sorted.bam"
bam = pysam.AlignmentFile(MT_BAM, "rb")

for read in bam.fetch():
    if read.is_unmapped:
        continue

    cluster_idx = read_cluster_map.get(read.query_name)
    if cluster_idx is not None:
        clusters.at[cluster_idx, "num_split_reads"] += 1

bam.close()

print(clusters[["ref_name", "bin_start", "num_reads", "num_split_reads"]].head())

Filters clusters again based on how many reads within each specific cluster align to the mt reference genome

OUTPUT_CSV = "../results/confirmed_NUMT_clusters.csv"

MIN_SPLIT_READS = 5 # How many split reads within a cluster must align with the mt genome to be counted

confirmed_clusters = clusters[clusters["num_split_reads"] >= MIN_SPLIT_READS]

print(f"Clusters before filtering: {len(clusters)}")
print(f"Clusters after filtering: {len(confirmed_clusters)}")

confirmed_clusters.to_csv(OUTPUT_CSV, index=False)
